More Vue structring, Fixed PFP, & Some bug fixes
