+++ src-tauri/tauri.conf.json
@@ -63,7 +63,7 @@
       },
       "protocol": {
         "asset": true,
-        "assetScope": ["$APPCONFIG/cache/games/banners/**"]
+        "assetScope": ["$APPCONFIG/cache/games/banners/**", "$APPCONFIG/cache/user/**"]
       },
       "window": {
         "center": true,
