+++ src/App.vue
@@ -3,7 +3,6 @@
   <titlebar-comp></titlebar-comp>
 
   <div class="bg" id="bg">
-    <div class="outline"></div>
     <div class="homebox" id="home">
       <div class="children fadeInUp">
         <img
@@ -34,78 +33,7 @@
       <div id="recentGamesList" class="fadeInDown gamesList"></div>
     </div>
 
-    <div class="secondorybox" id="games">
-      <div class="addGamePopUp" id="addGamePopUp">
-        <div class="mainSection fadeInDown">
-          <div class="section">
-            <div class="title">Game Name</div>
-            <input
-              maxlength="24"
-              type="text"
-              class="inputGameName"
-              id="inputGameName"
-            />
-          </div>
-          <div class="section">
-            <button class="addGameLocation" id="addGameLocation">
-              Locate Game
-            </button>
-          </div>
-          <div class="section">
-            <button class="addGameFinalBtn" id="addGameFinalBtn">
-              Add Game
-            </button>
-          </div>
-        </div>
-        <div class="addGameBannerSection fadeInUp">
-          <label for="addGameCustomBanner"></label>
-          <input
-            class="banner"
-            id="addGameCustomBanner"
-            type="file"
-            accept="image/png"
-            @change="(event) => loadCustomBanner(event)"
-          />
-          <p id="addGameCustomBannerTxt">Hover & Click to Select Banner</p>
-          <div
-            id="addGameCustomBannerOutput"
-            class="addGameCustomBannerOutput"
-          ></div>
-        </div>
-      </div>
-
-      <p>All Games</p>
-      <svg
-        id="addGameBtn"
-        class="addGameBtn"
-        version="1.1"
-        xmlns="http://www.w3.org/2000/svg"
-        xmlns:xlink="http://www.w3.org/1999/xlink"
-        x="0px"
-        y="0px"
-        width="459.325px"
-        height="459.325px"
-        viewBox="0 0 459.325 459.325"
-        style="enable-background: new 0 0 459.325 459.325"
-        xml:space="preserve"
-      >
-        <g>
-          <path
-            fill="var(--svgcolor)"
-            d="M459.319,229.668c0,22.201-17.992,40.193-40.205,40.193H269.85v149.271c0,22.207-17.998,40.199-40.196,40.193
-		c-11.101,0-21.149-4.492-28.416-11.763c-7.276-7.281-11.774-17.324-11.769-28.419l-0.006-149.288H40.181
-		c-11.094,0-21.134-4.492-28.416-11.774c-7.264-7.264-11.759-17.312-11.759-28.413C0,207.471,17.992,189.475,40.202,189.475h149.267
-		V40.202C189.469,17.998,207.471,0,229.671,0c22.192,0.006,40.178,17.986,40.19,40.187v149.288h149.282
-		C441.339,189.487,459.308,207.471,459.319,229.668z"
-          />
-        </g>
-      </svg>
-
-      <div class="search-bar">
-        <input type="text" placeholder="Search" id="gamesInput" />
-      </div>
-      <div id="allGamesList" class="fadeInDown gamesList"></div>
-    </div>
+    <allgames-comp></allgames-comp>
 
     <div class="secondorybox" id="favs">
       <p>Favourite Games</p>
@@ -134,172 +62,29 @@
       </div>
     </div>
 
-    <div class="leftbar">
-      <div class="d-flex justify-content-center">
-        <div>
-          <div class="user-pfp">
-            <label for="file"></label>
-            <input
-              id="file"
-              type="file"
-              accept="image/png"
-              @change="(event) => loadPFP(event)"
-            />
-            <img
-              src="./assets/svg/default-profile.svg"
-              alt="Avatar"
-              width="88"
-              id="output"
-            />
-          </div>
-          <input
-            class="username"
-            id="text"
-            type="text"
-            value="Lazap"
-            spellcheck="false"
-            maxlength="12"
-          />
-        </div>
-      </div>
-
-      <div class="side-tabs">
-        <div class="d-flex justify-content-center pb-2">
-          <div class="side-tab" id="home-btn">
-            <div class="indicator" id="indicator"></div>
-            <img src="./assets/svg/home.svg" height="25" width="25" />
-            <div class="side-tab-text">Home</div>
-          </div>
-        </div>
-
-        <div class="category-name d-flex justify-content-left">Games</div>
-        <div class="d-flex justify-content-center">
-          <div class="side-tab" id="recent-btn">
-            <img
-              id="recently-btn-img"
-              src="./assets/svg/recent.svg"
-              height="25"
-              width="25"
-            />
-            <div class="side-tab-text">Recent</div>
-          </div>
-        </div>
-        <div class="d-flex justify-content-center">
-          <div class="side-tab" id="games-btn">
-            <img
-              id="games-btn-img"
-              src="./assets/svg/games.svg"
-              height="25"
-              width="25"
-            />
-            <div class="side-tab-text">All Games</div>
-          </div>
-        </div>
-        <div class="d-flex justify-content-center pb-2">
-          <div class="side-tab" id="favs-btn">
-            <img
-              id="favs-btn-img"
-              src="./assets/svg/favs.svg"
-              height="25"
-              width="25"
-            />
-            <div class="side-tab-text">Favourites</div>
-          </div>
-        </div>
-
-        <div class="category-name d-flex justify-content-start">Friends</div>
-        <div class="d-flex justify-content-center">
-          <div class="side-tab" id="messages-btn">
-            <img src="./assets/svg/messages.svg" height="25" width="25" />
-            <div class="side-tab-text">Messages</div>
-          </div>
-        </div>
-        <div class="d-flex justify-content-center">
-          <div class="side-tab" id="activity-btn">
-            <img src="./assets/svg/activity.svg" height="25" width="25" />
-            <div class="side-tab-text">Activity</div>
-          </div>
-        </div>
-        <div class="d-flex justify-content-center">
-          <div class="side-tab" id="friends-btn">
-            <img src="./assets/svg/friends.svg" height="25" width="25" />
-            <div class="side-tab-text">All Friends</div>
-          </div>
-        </div>
-      </div>
-    </div>
+    <leftbar-comp></leftbar-comp>
   </div>
 </template>
 
-<script scoped>
+<script>
 import Settings from "./components/Settings.vue";
 import Titlebar from "./components/Titlebar.vue";
-import { selOption } from "./components/modules/rpcOptions";
+import AllGames from "./components/AllGames.vue";
+import LeftBar from "./components/LeftBar.vue";
+
+const path = window.__TAURI__.path;
+const invoke = window.__TAURI__.invoke;
 
 export default {
   name: "App",
   components: {
     "settings-comp": Settings,
     "titlebar-comp": Titlebar,
-  },
-  methods: {
-    async loadPFP(event) {
-      var selectedFile = event.target.files[0];
-      var reader = new FileReader();
-
-      reader.onload = async function () {
-        await window.__TAURI__.fs.writeBinaryFile(
-          (await window.__TAURI__.path.appDir()) + `storage/cache/user/pfp.png`,
-          reader.result
-        );
-        document.getElementById("output").src =
-          window.__TAURI__.tauri.convertFileSrc(
-            (await window.__TAURI__.path.appDir()) +
-              `storage/cache/user/pfp.png`
-          ) + `?${new Date().getSeconds()}`;
-      };
-      reader.readAsArrayBuffer(selectedFile);
-    },
-    async loadCustomBanner(event) {
-      var selectedFile = event.target.files[0];
-      var reader = new FileReader();
-
-      reader.onload = async function () {
-        await window.__TAURI__.fs.writeBinaryFile(
-          (await window.__TAURI__.path.appDir()) +
-            `storage/cache/games/banners/newcustombanner.png`,
-          reader.result
-        );
-        document.getElementById(
-          "addGameCustomBannerOutput"
-        ).style.backgroundImage =
-          `url(` +
-          window.__TAURI__.tauri.convertFileSrc(
-            (await window.__TAURI__.path.appDir()) +
-              `storage/cache/games/banners/newcustombanner.png`
-          ) +
-          `?${new Date().getSeconds()}` +
-          `)`;
-      };
-      document.getElementById("addGameCustomBannerTxt").style.opacity = "0";
-      reader.readAsArrayBuffer(selectedFile);
-    },
+    "allgames-comp": AllGames,
+    "leftbar-comp": LeftBar,
   },
   async mounted() {
     (async () => {
-      const fs = window.__TAURI__.fs;
-      const path = window.__TAURI__.path;
-      const dialog = window.__TAURI__.dialog;
-      const invoke = window.__TAURI__.invoke;
-
-      const home = document.getElementById("home");
-      const recent = document.getElementById("recent");
-      const games = document.getElementById("games");
-      const favs = document.getElementById("favs");
-      const friends = document.getElementById("friends");
-      const messages = document.getElementById("messages");
-      const activity = document.getElementById("activity");
-      const gameMenu = document.getElementById("gameMenu");
       const searchbars = document.querySelectorAll(
         'div.search-bar > input[type="text"]'
       );
@@ -318,25 +103,6 @@ export default {
   <div> <span style="margin-right: 4px;color:#5E81AC;">  </span> ${sysInfoInvoke.cpu}</div>`;
 
       await invoke("show_window");
-
-      try {
-        const data = JSON.parse(
-          await invoke("read_file", {
-            filePath: (await path.appDir()) + "cache/user/data.json",
-          })
-        );
-        document.getElementById("text").value = data.username;
-
-        await fs.readBinaryFile(
-          (await path.appDir()) + "storage/cache/user/pfp.png"
-        );
-        document.getElementById("output").src =
-          window.__TAURI__.tauri.convertFileSrc(
-            (await path.appDir()) + "storage/cache/user/pfp.png"
-          ) + `?${new Date().getSeconds()}`;
-      } catch (err) {
-        //console.log(err)
-      }
 
       try {
         let { accentColor } = JSON.parse(
@@ -396,177 +162,6 @@ export default {
         .loadGames("favGamesList", allGames, gamesdata)
         .catch((err) => {
           return console.error(err);
-        });
-
-      document
-        .getElementById("home-btn")
-        .addEventListener("click", async function () {
-          this.appendChild(document.getElementById("indicator"));
-          if (home.style.display !== "flex") {
-            toggleIndicatorAnim();
-          }
-
-          home.style.display = "flex";
-          recent.style.display = "none";
-          games.style.display = "none";
-          favs.style.display = "none";
-          friends.style.display = "none";
-          messages.style.display = "none";
-          activity.style.display = "none";
-          gameMenu.style.display = "none";
-
-          await require("./components/launchers/find-games.js")
-            .loadGames("recentGamesListMainPage")
-            .catch((err) => {
-              return console.error(err);
-            });
-
-          setActivity("home");
-        });
-
-      document
-        .getElementById("recent-btn")
-        .addEventListener("click", async function () {
-          this.appendChild(document.getElementById("indicator"));
-          if (recent.style.display !== "flex") {
-            toggleIndicatorAnim();
-          }
-
-          home.style.display = "none";
-          recent.style.display = "flex";
-          games.style.display = "none";
-          favs.style.display = "none";
-          friends.style.display = "none";
-          messages.style.display = "none";
-          activity.style.display = "none";
-          gameMenu.style.display = "none";
-
-          await require("./components/launchers/find-games.js")
-            .loadGames("recentGamesList")
-            .catch((err) => {
-              return console.error(err);
-            });
-
-          setActivity("recent");
-        });
-
-      document
-        .getElementById("games-btn")
-        .addEventListener("click", async function () {
-          this.appendChild(document.getElementById("indicator"));
-
-          if (games.style.display !== "flex") {
-            setTimeout(
-              () => document.getElementById("gamesInput").focus(),
-              100
-            );
-            toggleIndicatorAnim();
-          }
-
-          home.style.display = "none";
-          recent.style.display = "none";
-          games.style.display = "flex";
-          favs.style.display = "none";
-          friends.style.display = "none";
-          messages.style.display = "none";
-          activity.style.display = "none";
-          gameMenu.style.display = "none";
-
-          await require("./components/launchers/find-games.js")
-            .loadGames("allGamesList")
-            .catch((err) => {
-              return console.error(err);
-            });
-
-          setActivity("games");
-        });
-
-      document
-        .getElementById("favs-btn")
-        .addEventListener("click", async function () {
-          this.appendChild(document.getElementById("indicator"));
-
-          if (favs.style.display !== "flex") {
-            setTimeout(() => document.getElementById("favsInput").focus(), 100);
-            toggleIndicatorAnim();
-          }
-
-          home.style.display = "none";
-          recent.style.display = "none";
-          games.style.display = "none";
-          favs.style.display = "flex";
-          messages.style.display = "none";
-          activity.style.display = "none";
-          friends.style.display = "none";
-          gameMenu.style.display = "none";
-
-          await require("./components/launchers/find-games.js")
-            .loadGames("favGamesList")
-            .catch((err) => {
-              return console.error(err);
-            });
-
-          setActivity("favourites");
-        });
-
-      document
-        .getElementById("messages-btn")
-        .addEventListener("click", async function () {
-          this.appendChild(document.getElementById("indicator"));
-          if (messages.style.display !== "flex") {
-            toggleIndicatorAnim();
-          }
-
-          home.style.display = "none";
-          recent.style.display = "none";
-          games.style.display = "none";
-          favs.style.display = "none";
-          messages.style.display = "flex";
-          activity.style.display = "none";
-          friends.style.display = "none";
-          gameMenu.style.display = "none";
-
-          setActivity("messages");
-        });
-
-      document
-        .getElementById("activity-btn")
-        .addEventListener("click", async function () {
-          this.appendChild(document.getElementById("indicator"));
-          if (activity.style.display !== "flex") {
-            toggleIndicatorAnim();
-          }
-
-          home.style.display = "none";
-          recent.style.display = "none";
-          games.style.display = "none";
-          favs.style.display = "none";
-          messages.style.display = "none";
-          activity.style.display = "flex";
-          friends.style.display = "none";
-          gameMenu.style.display = "none";
-
-          setActivity("activity");
-        });
-
-      document
-        .getElementById("friends-btn")
-        .addEventListener("click", async function () {
-          this.appendChild(document.getElementById("indicator"));
-          if (friends.style.display !== "flex") {
-            toggleIndicatorAnim();
-          }
-
-          home.style.display = "none";
-          recent.style.display = "none";
-          games.style.display = "none";
-          favs.style.display = "none";
-          messages.style.display = "none";
-          activity.style.display = "none";
-          friends.style.display = "flex";
-          gameMenu.style.display = "none";
-
-          setActivity("friends");
         });
 
       document
@@ -630,134 +225,6 @@ export default {
       });
 
       document
-        .getElementById("addGameBtn")
-        .addEventListener("click", async function () {
-          if (
-            window.getComputedStyle(document.getElementById("addGamePopUp"))
-              .display === "flex"
-          ) {
-            document.getElementById(
-              "addGameCustomBannerOutput"
-            ).style.backgroundImage = "url()";
-            document.getElementById("addGamePopUp").style.display = "none";
-            document.getElementById("inputGameName").value = "";
-            document.getElementById("addGameCustomBannerTxt").style.opacity =
-              "1";
-            try {
-              await fs.removeFile(
-                (await path.appDir()) +
-                  `storage/cache/games/banners/newcustombanner.png`
-              );
-            } catch (e) {
-              return e;
-            }
-          } else {
-            document.getElementById("addGamePopUp").style.display = "flex";
-            document.getElementById("inputGameName").focus();
-          }
-        });
-
-      let newGameLocation;
-      document
-        .getElementById("addGameLocation")
-        .addEventListener("click", async function () {
-          const selected = await dialog.open({
-            multiple: false,
-            filters: [
-              {
-                name: "Windows Executable",
-                extensions: ["exe"],
-              },
-            ],
-          });
-          if (selected !== null) {
-            newGameLocation = selected;
-          }
-        });
-
-      document
-        .getElementById("addGameFinalBtn")
-        .addEventListener("click", async function () {
-          if (
-            document.getElementById("inputGameName").value.trim().length > 0 &&
-            newGameLocation
-          ) {
-            let scheme = {
-              DisplayName: document.getElementById("inputGameName").value,
-              LauncherName: "CustomGame",
-              GameID: "CustomGame",
-              Executable: newGameLocation.split("\\").slice(-1)[0],
-              Location: newGameLocation.split("\\").slice(0, -1).join("\\"),
-              Args: [],
-            };
-
-            try {
-              await fs.renameFile(
-                (await path.appDir()) +
-                  `storage/cache/games/banners/newcustombanner.png`,
-                (await path.appDir()) +
-                  `storage/cache/games/banners/${require("./components/modules/sha256.js").sha256(
-                    document
-                      .getElementById("inputGameName")
-                      .value.replaceAll(" ", "_")
-                  )}.png`
-              );
-              require("./components/launchers/find-games.js").Elements.createGameElement(
-                scheme,
-                "allGamesList"
-              );
-              let data = JSON.parse(
-                await invoke("read_file", {
-                  filePath: (await path.appDir()) + "cache/games/data.json",
-                })
-              );
-              data.push(scheme);
-              await invoke("write_file", {
-                filePath: (await path.appDir()) + "cache/games/data.json",
-                fileContent: JSON.stringify(data),
-              });
-            } catch (e) {
-              require("./components/launchers/find-games.js").Elements.createGameElement(
-                scheme,
-                "allGamesList"
-              );
-              let data = JSON.parse(
-                await invoke("read_file", {
-                  filePath: (await path.appDir()) + "cache/games/data.json",
-                })
-              );
-              data.push(scheme);
-              await invoke("write_file", {
-                filePath: (await path.appDir()) + "cache/games/data.json",
-                fileContent: JSON.stringify(data),
-              });
-            }
-            document.getElementById(
-              "addGameCustomBannerOutput"
-            ).style.backgroundImage = "url()";
-            document.getElementById("addGamePopUp").style.display = "none";
-            document.getElementById("inputGameName").value = "";
-            document.getElementById("addGameCustomBannerTxt").style.opacity =
-              "1";
-            newGameLocation = "";
-          } else if (
-            document.getElementById("inputGameName").value.trim().length > 0 &&
-            !newGameLocation
-          ) {
-            return alert("You are missing the game location.");
-          } else if (
-            document.getElementById("inputGameName").value.length < 0 &&
-            newGameLocation
-          ) {
-            return alert("You are missing the game name.");
-          } else {
-            return alert(
-              "You are missing the game name, and the game location."
-            );
-          }
-        });
-
-      document
         .getElementById("removeGame")
         .addEventListener("click", async function () {
           let data = JSON.parse(
@@ -794,13 +261,6 @@ export default {
           );
         });
 
-      function toggleIndicatorAnim() {
-        let indicator = document.getElementById("indicator");
-        indicator.classList.add("anim-indicatorscaleY");
-        setTimeout(() => {
-          indicator.classList.remove("anim-indicatorscaleY");
-        }, 200);
-      }
       function updateAccentColor(accentColor) {
         document.getElementById("indicator").style.backgroundColor =
           accentColor;
@@ -810,7 +270,7 @@ export default {
       }
       async function setActivity(tab) {
         const { state, details, largeImage, largeText, smallImage, smallText } =
-          selOption(tab);
+          require("./components/modules/rpcOptions").selectOption(tab);
         if (timestamp === null) timestamp = Date.now();
         try {
           await invoke(`set_activity`, {
@@ -853,7 +313,7 @@ export default {
   src: url("./assets/fonts/Nunito-SemiBold.ttf") format("truetype");
 }
 
-html, 
+html,
 body {
   zoom: 0.945;
   background: #181a1f;
@@ -1174,164 +634,6 @@ body {
 
 .rightbar .stickers img {
   width: 100px;
-}
-
-.leftbar {
-  background-color: #1c1d22;
-  border-radius: 15px;
-  position: absolute;
-  height: calc(100% - 71px);
-  max-height: 100%;
-  width: 270px;
-  top: 25px;
-  margin-top: 30px;
-  margin-right: 100%;
-  margin-left: 300px;
-  text-align: center;
-  box-shadow: 0 3px 16px -7px rgb(17 18 24 / 70%);
-}
-
-.user-pfp {
-  position: relative;
-  margin-top: 50px;
-  width: 88px;
-  height: 88px;
-  margin-left: auto;
-  margin-right: auto;
-  transition: all 0.2s ease-in-out;
-}
-
-.user-pfp:hover {
-  cursor: pointer;
-  transition: all 0.2s ease-in-out;
-  border-radius: 14px;
-  scale: 1.1;
-}
-
-.user-pfp input {
-  display: none;
-}
-
-.user-pfp img {
-  border-radius: 18px;
-  image-rendering: auto;
-  width: 88px;
-  height: 88px;
-  object-fit: cover;
-  box-shadow: 0 3px 16px -7px rgb(17 18 24 / 70%);
-}
-
-.user-pfp label {
-  position: absolute;
-  cursor: pointer;
-  height: 70px;
-  width: 70px;
-}
-
-.username {
-  margin-top: 8px;
-  margin-bottom: 26px;
-  text-align: center;
-  font-size: 22px;
-  font-family: Nunito-Bold;
-  color: #b9b9b9;
-  background-color: transparent;
-  border: transparent;
-  text-decoration: none;
-}
-
-.side-tabs {
-  margin-left: 20px;
-  position: absolute;
-  height: 66%;
-  overflow-y: auto;
-}
-
-.side-tab {
-  background-color: rgba(30, 30, 34, 0.5);
-  box-shadow: -1px 3px 8px -1px rgba(0, 0, 0, 0.2);
-  height: 40px;
-  width: 230px;
-  border-radius: 10px;
-  display: flex;
-  margin-bottom: 10px;
-  z-index: 1;
-  transition: color 0.15s ease-out;
-}
-
-.side-tab:before {
-  background: rgb(30, 30, 34);
-  background: linear-gradient(
-    30deg,
-    rgba(30, 30, 34, 1) 0%,
-    rgba(30, 30, 34, 1) 30%,
-    var(--back) 100%
-  );
-  content: "";
-  height: 40px;
-  width: 230px;
-  border-radius: 10px;
-  display: flex;
-  margin-bottom: 10px;
-  position: absolute;
-  z-index: -1;
-  opacity: 0;
-  transition: opacity 0.15s ease-out;
-}
-
-.side-tab:hover:before {
-  opacity: 1;
-}
-
-.side-tab:hover {
-  cursor: default;
-  color: #fff;
-}
-
-.side-tab img {
-  filter: invert(100%) sepia(6%) saturate(0%) hue-rotate(115deg)
-    brightness(108%) contrast(108%);
-  margin-left: 25px;
-  margin-top: 8px;
-  margin-right: 5px;
-  display: inline-block;
-}
-
-.side-tab .side-tab-text {
-  color: #dadada;
-  font-size: 18px;
-  margin-left: 16px;
-  align-self: center;
-  margin-top: 4px;
-  cursor: default;
-}
-
-.category-name {
-  color: #7a7a7a;
-  margin-left: 10px;
-  margin-bottom: 4px;
-  font-size: 15px;
-}
-
-.category-name:hover {
-  cursor: default;
-}
-
-.indicator {
-  transition: 0.3s;
-  position: absolute;
-  margin-top: 5px;
-  margin-left: 2px;
-
-  background-color: var(--back);
-  height: 32px;
-  width: 5px;
-  border-radius: 50px;
-  display: inline;
-}
-
-.anim-indicatorscaleY {
-  animation: indicatorscaleY 0.2s linear;
 }
 
 .gamesList {
@@ -1657,7 +959,6 @@ img,
   justify-content: space-between;
   align-items: center;
 }
-
 
 .leftbar-overlay {
   position: absolute;
