+++ src/components/Settings.vue
@@ -99,8 +99,6 @@
 </template>
 
 <script>
-import { selOption } from "./modules/rpcOptions";
-
 export default {
   name: "settings-comp",
   async mounted() {
@@ -228,7 +226,7 @@ export default {
     async function setActivity(tab) {
       let timestamp = null;
       const { state, details, largeImage, largeText, smallImage, smallText } =
-        selOption(tab);
+        require("./modules/rpcOptions").selectOption(tab);
       if (timestamp === null) timestamp = Date.now();
       try {
         await invoke(`set_activity`, {
